# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog Laravel OpenApi",
        "description": "Blog Laravel Swagger OpenApi",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "3.1.0"
    },
    "paths": {
        "/api/admin/posts/{postId}/comments/{commentId}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get Comment For Admin",
                "description": "Get Comment For Admin",
                "operationId": "GetCommentForAdmin",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of Post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID of Comment",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete Comment By Admin",
                "description": "Delete Comment By Admin",
                "operationId": "DeleteCommentByAdmin",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of Post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID of Comment",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/posts/{postId}/comments/{commentId}/restore": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Restore Comment By Admin",
                "description": "Restore Comment By Admin",
                "operationId": "RestoreCommentByAdmin",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of Post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID of Comment",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restore",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/posts/{postId}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get All Comments",
                "description": "Get All Comments",
                "operationId": "GetAllComments",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of Post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Index",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create Comment",
                "description": "Create Comment",
                "operationId": "CreateComment",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of Post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                ],
                "requestBody": {
                    "description": "create a comment",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "text": "example comment"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{postId}/comments/{commentId}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get Comment",
                "description": "Get Comment",
                "operationId": "GetComment",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of Post",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID of Comment",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}